Refactor the PubNub client into smaller modules
